requirements.txt
qrcode>=7.4.2
Pillow>=9.5.0